fix(sync): keep profile rollout tests protected
